+++ changelog.md
@@ -1,2 +1,3 @@
 - Fix options list not resizing correctly on screen switch
 - Fix options reset screens
+- Add unicode support to trigger check